// src/app/features/transfer-monitor/models/conversion-monitoring.model.ts
export interface ConversionMonitoring {
  BhNo: string;
  BhName: string;
  FCode: string;
  DDate: string;
  Status: string;
  TrStatus: string;
  TrDate: string;
  TrTime: string;
  Times: number;
  ReadCt: number;
  NewCt: number;
  OldCt: number;
  DeleteCt: number;
  TrUser: string;
  FolderPath: string;
  Descript: string;
}